fix(sdk): do not export default payload in as-code (#4117)
close #4116
